POC: convert function checks for units
Since there is only one convert function, it's easy to apply different
convertion strategy for specific units like temperature that requires a
formula, not a simple coefficient. It's will be harder for inverse
conversion, because there is no way to invert a function. Or is there?

At least I know that amount of formulae is less than all combinations,
because transition works here the same way as for coefficients - I used
resursion calls few times.